+++ source/dotnet/Services/Application/ProcessResult/IProcessStepResultFactory.cs
@@ -13,12 +13,12 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Contracts;
-using Energinet.DataHub.Wholesale.Domain.ProcessOutput;
+using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
 using Energinet.DataHub.Wholesale.Domain.ProcessStepResultAggregate;
 
 namespace Energinet.DataHub.Wholesale.Application.ProcessResult;
 
-public interface IProcessStepResultMapper
+public interface IProcessStepResultFactory
 {
-    ProcessStepResultDto MapToDto(ProcessStepResult processStepResult);
+    ProcessStepResultDto Create(ProcessStepResult processStepResult, Batch batchDto);
 }