+++ source/dotnet/Services/Application/ProcessResult/ProcessStepResultApplicationService.cs
@@ -13,6 +13,7 @@
 // limitations under the License.
 
 using Energinet.DataHub.Wholesale.Contracts;
+using Energinet.DataHub.Wholesale.Domain.BatchAggregate;
 using Energinet.DataHub.Wholesale.Domain.GridAreaAggregate;
 using Energinet.DataHub.Wholesale.Domain.ProcessStepResultAggregate;
 
@@ -24,21 +25,23 @@ namespace Energinet.DataHub.Wholesale.Application.ProcessResult;
 public class ProcessStepResultApplicationService : IProcessStepResultApplicationService
 {
     private readonly IProcessStepResultRepository _processStepResultRepository;
-    private readonly IProcessStepResultMapper _processStepResultMapper;
+    private readonly IProcessStepResultFactory _processStepResultFactory;
+    private readonly IBatchRepository _batchRepository;
 
-    public ProcessStepResultApplicationService(IProcessStepResultRepository processStepResultRepository, IProcessStepResultMapper processStepResultMapper)
+    public ProcessStepResultApplicationService(IProcessStepResultRepository processStepResultRepository, IProcessStepResultFactory processStepResultFactory, IBatchRepository batchRepository)
     {
         _processStepResultRepository = processStepResultRepository;
-        _processStepResultMapper = processStepResultMapper;
+        _processStepResultFactory = processStepResultFactory;
+        _batchRepository = batchRepository;
     }
 
     public async Task<ProcessStepResultDto> GetResultAsync(ProcessStepResultRequestDto processStepResultRequestDto)
     {
-        var processActorResult = await _processStepResultRepository.GetAsync(
+        var processStepResult = await _processStepResultRepository.GetAsync(
                 processStepResultRequestDto.BatchId,
                 new GridAreaCode(processStepResultRequestDto.GridAreaCode))
             .ConfigureAwait(false);
-
-        return _processStepResultMapper.MapToDto(processActorResult);
+        var batch = await _batchRepository.GetAsync(processStepResultRequestDto.BatchId).ConfigureAwait(false);
+        return _processStepResultFactory.Create(processStepResult, batch);
     }
 }