+++ source/dotnet/Services/WebApi/Configuration/ServiceCollectionExtensions.cs
@@ -86,7 +86,7 @@ internal static class ServiceCollectionExtensions
         services.AddScoped<ICalculatorJobRunner>(_ => null!); // Unused in the use cases of this app
         services.AddScoped<ICalculatorJobParametersFactory>(_ => null!); // Unused in the use cases of this app
         services.AddScoped<IProcessStepResultApplicationService, ProcessStepResultApplicationService>();
-        services.AddScoped<IProcessStepResultMapper, ProcessStepResultMapper>();
+        services.AddScoped<IProcessStepResultFactory, ProcessStepResultFactory>();
         services.AddScoped<IProcessStepResultRepository, ProcessStepResultRepository>();
         services.AddScoped<IBatchRequestDtoValidator, BatchRequestDtoValidator>();