+++ source/dotnet/Packages/Contracts.Documentation/release-notes/release-notes.md
@@ -1,5 +1,9 @@
 # Wholesale Contracts Release notes
 
+## Version 1.0.4
+
+Add `PeriodStart` and `PeriodEnd` to `ProcessStepResultDto`
+
 ## Version 1.0.3
 
 Bump version as part of pipeline change.